api/requirements.txt
fastapi[all]
pymongo
firebase-admin
python-dotenv
redis
aioredis